Send signal when suspend is canceled.
Send a PowerStateChanged "on" signal when suspend is canceled. This is
the same signal sent by powerd_suspend when resuming from suspend. This
will tell devices that have received the suspend signal to proceed
normally.

This patch is cherry-picked from http://codereview.chromium.org/5974008
and modified accordingly.

BUG=chromium-os:33852
TEST=Verify with dbus-monitor that a PowerStateChanged:on DBus signal is
sent when the suspend is canceled.

Change-Id: I2299ddd3555559a01afa2ef2cf20ff3d1254b973
Reviewed-on: https://gerrit.chromium.org/gerrit/36540
Commit-Ready: Ben Chan <benchan@chromium.org>
Tested-by: Ben Chan <benchan@chromium.org>
Reviewed-by: Ben Chan <benchan@chromium.org>