[login_manager] Add 'liveness' pings to detect and mitigate browser hangs.
In order to detect browser hangs and generate actionable crash reports,
have the session manager send 'liveness' pings to chrome over DBus.  If
it's time to send a new ping and Chrome hasn't responded, the SM will SIGABRT
it to generate a crash dump.

We considered using a different signal to make these obvious, but were unable
to come up with something that wouldn't be problematic.

BUG=chromium-os:35729
TEST=unit
TEST=http://chromegw/i/chromiumos.tryserver/builders/lumpy-paladin/builds/384

Change-Id: I944ffc4d4d33179df710ffb5a8617c3d365dbc06
Reviewed-on: https://gerrit.chromium.org/gerrit/37864
Tested-by: Chris Masone <cmasone@chromium.org>
Reviewed-by: Chris Masone <cmasone@chromium.org>
Commit-Ready: Chris Masone <cmasone@chromium.org>