login: Copy host background if login animations are disabled
Pass --ash-copy-host-background-at-boot to Chrome if the
disable_login_animations USE flag is set.  This makes Chrome
copy the boot splash screen to a compositor layer so it can
be displayed while we're loading the login background image;
otherwise, a blank white screen is shown for several
seconds.

BUG=chromium:152309
TEST=manual: no blank white screen between boot splash
     screen and login background on x86-mario

Change-Id: I21b42e568c4fa9e420866657c5bf915d79e84604
Reviewed-on: https://gerrit.chromium.org/gerrit/36949
Tested-by: Daniel Erat <derat@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Commit-Ready: Daniel Erat <derat@chromium.org>