Change threshold powerd rounds up to 100% when no current to 98%
When there is no current going into the battery and the battery is near full
this normally means that the controller is either no charging the battery or
trickle charging it. Thus leaving it charge for longer will no prolong the
battery life when off AC power. From a user perspective this means that the
battery is full, so we adjust the displayed percentage. This changes the
threshold for this from 99% to 98% to catch more occurences of this.

BUG=chromium-os:35818
TEST=This is just changing a constant for a case that is non-trivial to produce,
     so there is no test other then confirming the package builds and unit tests
     pass. The functional details of this would have been tested when the
     original code was written.

Change-Id: I6183bbd73bc49c6e5a3696a5648169dd029f26f3
Reviewed-on: https://gerrit.chromium.org/gerrit/36925
Reviewed-by: Ryan Harrison <rharrison@chromium.org>
Tested-by: Ryan Harrison <rharrison@chromium.org>
Commit-Ready: Ryan Harrison <rharrison@chromium.org>