Fill free memory with pattern and verify on resume
This CL adds memory_suspend_test, a wrapper around powerd_suspend that
will allocate memory and fill it with a known pattern to verify on resume.
It allocates available memory (available + inactive) and subtracts 100 MB
to leave slack space to prevent the oom killer from killing processes
during suspend.

This is intended to verify that memory does not get corrupted during the
suspend/resume process and that kernel checks for this condition work
properly.

Additionally cleanup was made to suspend_stress_test logic flows to
support the changes.  It will notice multiple errors in the loop
and print out all of them before exiting if a failure is detected.

BUG=chrome-os-partner:15654
TEST=Force failures for suspend_stress_test, validate loop exits
  Force multiple failures, ensure all errors are printed
  Ran on machine with bad bios, saw memory errors reported from
    memory_suspend_test matched CRC errors reported in kernel.

Change-Id: Ia49df9758e18824fb1995e1ad2669fec88d7396d
Reviewed-on: https://gerrit.chromium.org/gerrit/36944
Tested-by: Jon Kliegman <kliegs@chromium.org>
Reviewed-by: Julius Werner <jwerner@chromium.org>
Commit-Ready: Jon Kliegman <kliegs@chromium.org>