Pin power to full for 3 minutes after leaving AC
Since we round up when we are near full, when we come out of being on AC there
is the possibility of their being a significant drop in the battery percentage,
which can be disconcerting to a user. To avoid this issue we are now pinning
value to full for 3 minutes after leaving AC and then tapering from the pinned
to the hardware reported value over a period of 7 minutes. This required a
complete refactoring of the battery percentage code in powerd, since there was
already a linear scaling being done to the data.

BUG=chromium-os:35819
TEST=Build and run unit tests.
     With charged battery pulled AC and confirmed battery percent stayed @ 100%
     for 3 minutes.
     After 3 minutes confirmed that there was no large drop in value (ex. 100%
     -> 98%) over the next 7 minutes.

Change-Id: I17637110735d82286aeb771a9ce096767aea5422
Reviewed-on: https://gerrit.chromium.org/gerrit/37053
Tested-by: Ryan Harrison <rharrison@chromium.org>
Commit-Ready: Ryan Harrison <rharrison@chromium.org>
Reviewed-by: Ryan Harrison <rharrison@chromium.org>