Add a stale bit for power status in powerd
This resolves a race condition in the Suspend/Resume flow where we send a signal
to Chrome that there is new power information before the suspend, but Chrome
doesn't request the new data until after the resume. We are marking the power
status as stale or dirty, so that if there are any requests for it, the DBus
handler polls for new information. The polling mechanism will clear the bit once
it restarts.

BUG=chromium:157144
TEST=Built and ran unittests.
     Built image and looked for regressions in the Suspend/Resume behaviour of
     the battery icon.
     Built image with additional logging and confirmed that the stale bit was
     being set and cleared when expected.

Change-Id: I0dfb2fb7a99f8e400e99df0ed4af189d66a19550
Reviewed-on: https://gerrit.chromium.org/gerrit/37457
Reviewed-by: Ryan Harrison <rharrison@chromium.org>
Tested-by: Ryan Harrison <rharrison@chromium.org>
Commit-Ready: Ryan Harrison <rharrison@chromium.org>