Enforce disable_als flag in AmbientLightSensor
Even though we have a disable_als pref, it is not being used to block
enabling/disabling of ALS.  This patch saves that pref as a member flag
and adds a check to EnableOrDisableSensor() to prevent ALS from being
enabled if that flag is set.

BUG=chromium-os:36126
TEST=echo 1 > /var/lib/power_manager/disable_als, restart powerd.
Wait for system to go to idle dim/screen-off.  Now go back to active
state.  See powerd.LATEST, make sure you see "ALS has been completely
disabled" and not "Enabling light sensor poll."

Change-Id: Iad0e14b625275d9577b907802db107f8eb418805
Signed-off-by: Simon Que <sque@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/37537
Reviewed-by: Daniel Erat <derat@chromium.org>