Allow backlight to stay on with active audio with keep_backlight_on_for_audio
On some systems the backlight is tied to the audio system of the device,
for such systems we need to keep the backlight on if audio is playing. This
enables systems to have the keep_backlight_on_for_audio flag to keep an active
backlight when the system is generating audio.

BUG=chrome-os-partner:13240
TEST=Manual, verify backlight stays on when the keep_backlight_on_for_audio
flag is present.

Change-Id: I2d1c71b4145c60182e2b63bc6e6c96d155b417bc
Reviewed-on: https://gerrit.chromium.org/gerrit/37584
Tested-by: Bernie Thompson <bhthompson@chromium.org>
Reviewed-by: Daniel Erat <derat@chromium.org>
Commit-Ready: Bernie Thompson <bhthompson@chromium.org>