shill: manager: Suppress autoconnect while stopped
The manager may continue to run for a short while after Manager::Stop()
is called, as shutdown tasks are executed.  Prevent the the AutoConnect
task from running as this could be triggered as Services are disconnected.
From there things can only get worse as the associated services for these
devices could be in a partially removed state and cause untold mahyem.

While here, print a message at the point where shill receives a
signal and is shutting down.

BUG=chromium-os:36149
TEST=Unit tests, kill shill with SIGTERM and make sure it shuts down correctly

Change-Id: Ifb7ae1ba4114bc696958498c15b1bf082d2d0e70
Reviewed-on: https://gerrit.chromium.org/gerrit/37581
Reviewed-by: Christopher Wiley <wiley@chromium.org>
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Commit-Ready: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>