Change user control of keyboard backlight
The user control is no longer a toggle on/off. Now it causes the backlight to
switch over to using user brightness levels and the up/down causes the light
move up or down a level. The light will stay pinned at the user selected level
until end of session. If the device idles/suspends then the backlight will be
temporarily set to a dim or off level, but will return to the previously defined
user level.

BUG=chromium-os:35376
TEST=Build and run unittests
     Confirm that ALS control of the backlight behaves as expected.
     Confirm that Alt+BrightnessUp/Down transitions to user control.
     Confirm that there is the expected number of brightness levels in user mode
     and they look smooth.
     Confirm that changes in the ambient light level do no cause the device to
     switch back to ALS control.
     Confirm that if the device idles or suspends when in user mode the
     backlight adjust appropriately, but returns after going active again.

Change-Id: I6b62d613a0f0808c6fc8284cbab86e2eed524a61
Reviewed-on: https://gerrit.chromium.org/gerrit/38305
Commit-Ready: Ryan Harrison <rharrison@chromium.org>
Reviewed-by: Ryan Harrison <rharrison@chromium.org>
Tested-by: Ryan Harrison <rharrison@chromium.org>