power: Simplify ambient light sensor code.
I'm removing the logic that turns ambient light sensor
polling on or off depending on whether the backlight is on
or not.  It added a fair amount of complexity and I don't
see much benefit -- polling just means that we read a file
from sysfs once a second, and we were already doing this the
majority of the time anyway.

I'm also moving the enforcement of the "disable_als" pref
from the AmbientLightSensor class into the
BacklightController implementations that consume the
readings.  There was already a TODO saying that we want to
be able to poll without making adjustments.

Also included is some miscellaneous cleanup to the prefs
code that I ran into while doing this.

Finally, AmbientLightSensor's stats-dumping methods claimed
to list the most recent values first, but actually listed
the oldest first, so I'm fixing that too.

BUG=chromium-os:36284
TEST=unit tests pass; internal display's brightness is
     adjusted in response to ambient brightness changes on
     lumpy; keyboard backlights work too; no backlight
     adjustments after setting disable_als pref to 1

Change-Id: Ic9c338cddcf5550288ceb7ecd26350097f9545bc
Reviewed-on: https://gerrit.chromium.org/gerrit/37854
Commit-Ready: Daniel Erat <derat@chromium.org>
Reviewed-by: Daniel Erat <derat@chromium.org>
Tested-by: Daniel Erat <derat@chromium.org>