shill: Disconnect modem before attempting to activate
When we're asked to activate a modem, assume that Chrome knows what it
is doing and disconnect the modem first.  This lets Chrome not have
complicate the activation process by waiting for an asyncronous
disconnect first.

BUG=chromium-os:36301
TEST=Modems still activate as before.  This shouldn't be surprising
because Chrome is still doing lots of work to disconnect a service
before trying to activate.  The real magic will happen when we take
those parts out of Chrome.

Change-Id: Ic41fa7f4ec759d6ec5c650708792ab6aa083dde5
Reviewed-on: https://gerrit.chromium.org/gerrit/38060
Commit-Ready: Christopher Wiley <wiley@chromium.org>
Reviewed-by: Christopher Wiley <wiley@chromium.org>
Tested-by: Christopher Wiley <wiley@chromium.org>