shill: Add Nl80211Attribute type.
This checkin is just a framework for the Nl80211Attribute type.
The Nl80211Attribute type encapsulates each attribute in an object.
This will enable the creation of an AttributeList type that can
encapsulate an arbitrary number of levels of nested attributes.

The next checkin includes all the types and all the attributes
necessary to use Nl80211Attribute.  It also passes unittests.

BUG=chromium-os:36637
TEST=unittests.

Change-Id: Ide967812e5286526d37f3ac5d478ab5a4c358bd6
Reviewed-on: https://gerrit.chromium.org/gerrit/38973
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Commit-Ready: Wade Guthrie <wdg@chromium.org>
Tested-by: Wade Guthrie <wdg@chromium.org>