shill: Don't let WiFi scan during association or configuration
Don't let the UI trigger scans during association or configuration to
avoid missing frames while we're off channel.  This fixes behavior where
we take a long time to connect to service (or fail completely) because
we essentially stop listening to the AP for a second or two.

BUG=chromium-os:36233
TEST=Unit tests

Change-Id: I5cd9c687209b13d6206930f55761ba355201ff82
Reviewed-on: https://gerrit.chromium.org/gerrit/38482
Commit-Ready: Christopher Wiley <wiley@chromium.org>
Reviewed-by: Christopher Wiley <wiley@chromium.org>
Tested-by: Christopher Wiley <wiley@chromium.org>