Trigger saving RTC resume timestamp by file existance instead of flag
powerd_suspend can be told to save an RTC timestamp as soon as possible
during resume by passing the -t flag, which helps the power_Resume
autotest get more accurate timing measurements. More recently, the
power_UiResume test was added to cover more code by triggering suspend
with a DBus message, which makes passing flags to powerd_suspend
difficult.

As both autotests are being refactored to share more code, it makes
sense to add the RTC timestamp measurement to power_UiResume too. To
facilitate that, this patch changes the powerd_suspend interface to
remove the -t flag and instead use the existance of the
hwclock-on-resume file as trigger for reading a timestamp on resume.
This allows both autotests to touch the file before suspend and use the
suspension method of their choice without a need to pass command line
flags.

BUG=chromium-os:35901
TEST=None

Change-Id: I8452eefd00866bc6ca6eb6164c810e4156682afb
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/38699
Reviewed-by: Sameer Nanda <snanda@chromium.org>