Pick last instead of first URL in v3 clients for now.
We have a design underway for trying out all the URLs in the Omaha rules.
But update_engine has already been modified to be v3 and pick the first
URL. What this means is that any bits released until the full design gets
implemented, in future, when we actually populate multiple URLs such as
HTTP and HTTPs, will pick HTTP and fail because it doesn't have the full
support to process HTTP URLs.

So we need to change that logic to always pick the last one. This is
based on the assumption that the last one will be more stable (e.g. HTTPS).

BUG=chromium-os:36724
TEST=Tested on ZGB with devserver by supplying multiple URLs.

Change-Id: I85291dae56c9be83eb83955d6e123cdb8e33f28a
Reviewed-on: https://gerrit.chromium.org/gerrit/38827
Reviewed-by: Don Garrett <dgarrett@chromium.org>
Commit-Ready: Jay Srinivasan <jaysri@chromium.org>
Tested-by: Jay Srinivasan <jaysri@chromium.org>