[init] rewrite date-proxy-watcher
date-proxy-watcher had a severe race condition, of this form:
1) request info
2) wait for response signal

Unfortunately if the response signal comes between 1 and 2 we hang forever. The
solution is to use a method of waiting that will buffer for us - i.e., a fifo.
Now this code uses three cooperating processes:

1) A process that waits for shill PropertyChange events, and fires off a request
for proxy info (1 above)
2) A process that waits for proxy info responses, and writes the info to a fifo
called "proxy-fifo" (2 above)
3) A process that blocks reading the fifo, figures out whether the proxy
actually changed, and notifies the parent process as needed.

BUG=chromium-os:31505
TEST=none yet

Change-Id: I47d11dbc1cc32c35dba34bbe1dd5d4973e1672ab
Signed-off-by: Elly Fong-Jones <ellyjones@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/38854
Reviewed-by: Richard Barnette <jrbarnette@chromium.org>