Salsa: Adding in a try_touch_experiment option to crosh
To facilitate the Salsa Touch testing framework, this is adding a new
command to Crosh.  This allows you to pass it an obfuscated command that
encodes a number of Touch Properties to change.  It will apply the
changes, but back up your original values.

The user can then, try out the new settings (and not know what they
changed since the values are hidden) and press Ctrl+C whenever they want
to revert.

So when a user is taking a Touch experiment, they will be given a big
command to copy/paste into Crosh and it will give them a temporary
experience without changing their own, personal settings.

BUG=chromium-os:36840
TEST=Run it with a sample input and confirm that the properties get
changed/revert correctly.  Try it with some invalid inputs and make sure
the users settings don't get messed up by it.

Change-Id: I6e1040d4a8e0e79868a579c11fe9266bdf2f0a67
Signed-off-by: Charlie Mooney <charliemooney@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/39136
Reviewed-by: Andrew de los Reyes <adlr@chromium.org>