[init] tlsdate: skip forward if needed
On November 13, we rolled a build image; 8 days later, the certificate for
clients3.google.com was replaced, and machines with that image on them (which
were defaulting to November 13) ceased to be able to fetch time. This is going
to happen basically whenever someone tries to use a sufficiently old build, and
to make matters worse, the certs on clients3.google.com are short-lived, so this
will bite us over and over.

Fortunately, there's a solution: 'leap mode'. In this mode, tlsdate will treat
the current time as though it was the start time on the received certificate
_if_ the start time passes normal time sanity checks and is after the current
time.

BUG=chromium-os:36738
TEST=adhoc
Run tlsdated without a cached time, check that the tlsdate subprocess runs with
-l using 'ps aux | grep tlsdate'.

Change-Id: Ib927cbe7370486f2cf104d4387df7f4a59d83542
Signed-off-by: Elly Fong-Jones <ellyjones@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/39107
Reviewed-by: Will Drewry <wad@chromium.org>