shill: Add type-specific children of Nl80211Attribute type.
This checkin adds all the type-specific children of the Nl80211Attribute
type without actually changing any of the code that's called (that's
planned for a later checkin).

The next checkin includes all the attributes necessary to use
Nl80211Attribute.  After that, we'll go for the functionality changes
(which, BTW, passes unittests).

BUG=chromium-os:36637
TEST=unittests.

Change-Id: Ib3944e3b18faeec322ea93cd989daa9bbeb0c0c7
Reviewed-on: https://gerrit.chromium.org/gerrit/39319
Commit-Ready: Wade Guthrie <wdg@chromium.org>
Reviewed-by: Wade Guthrie <wdg@chromium.org>
Tested-by: Wade Guthrie <wdg@chromium.org>