shill: WiFi: Support cached credentials
Adopt a more "lazy" approach to removing wpa_supplicant network
entries for EAP-TLS networks, so that credential information can
be cached.  This speeds up the re-authentication process for some
network types.  To do this, we disable the network entries instead
of removing them from the wpa_supplicant interface.  To re-connect
we simply re-select the saved network block.

If any credentials are changed in the WiFi service, these cached
credentials should be removed, with the side effect of disconnecting
the network.

As a part of this change, we remove the now deprecated
ClearCachedCredentials wpa_supplicant interface API function which
is a since-reverted local modification.

BUG=chromium-os:25670,chrome-os-partner:15005
TEST=New unit tests.  Manually tested using a pair of scenarios
where there is both a user-profile-hosted network and a
default-profile-hosted network.  The two test scenarios begin by
establishing connections to both networks.  In one scenario,
leave the user-profile network connected.  In the other scenario
leave the default-profile network connected.  Use the "list"
command in wpa_cli to monitor what network entries are present
before and after logout.

In both cases, before logout, ensure that the connected network is
marked "ACTIVE" and the other network is marked "DISABLED",
indicating that wpa_supplicant will not automatically switch to
the un-selected network.

In the first scenario (user-profile network connected), after
logout, the user-profile network disappears.  The default-profile
network loses the "DISABLED" flag, as this network is now being
actively connected to.  Shortly after, it is labeled as "ACTIVE".

In the second scenario (default-profile network connected),
after logout the user-profile network disappears.  The default-profile
network remains ACTIVE and is not disconnected.

New autotest under development:
network_WiFiRoaming/network_WiFiRoaming.011PMKSA_Caching, which
tests that on re-association, the EAP-TLS step is bypassed, and
RSN PMKSA caching is used instead.  Tested using packet-capture
that an PMKID appears in the RSN IE in the association request.

Verified during the test using the wpa_cli "pmksa" command that
two PMKIDs appear, one for each of the APs used in the test, and
that at test completion they are both removed.

Change-Id: I0c45662716dcc35c89f1277c4d7949909003e13a
Reviewed-on: https://gerrit.chromium.org/gerrit/39676
Reviewed-by: mukesh agrawal <quiche@chromium.org>
Commit-Ready: Paul Stewart <pstew@chromium.org>
Tested-by: Paul Stewart <pstew@chromium.org>