Prevent powerd_suspend's hwclock call from causing aborted suspends
When run in test mode (formerly via -t flag, recently due to existence
of /var/run/power_manager/hwclock-on-resume), powerd_suspend calls
hwclock right before suspend to ensure it is cached on wakeup.

Due to a recent kernel change (gerrit/38480), the kernel interprets all
RTC interrupts as a wakeup reason. hwclock has to temporarily activate
periodical interrupts on the RTC to get an accurate sub-second reading,
which will then cause the kernel to abort immediately upon starting to
suspend. This has been reliably breaking all power_UiResume iterations
in the last two weeks, and would break power_Resume too when I am soon
going to change it to support aborted suspends.

This patch changes the hwclock call in powerd_suspend to use the -V
flag, which causes it to only print version information and exit. This
maintains the intended effect of running the program (caching the binary
and all its shared objects) but prevents it from making an actual RTC
access.

BUG=chromium-os:37154
TEST=Run power_UiResume, watch it not fail horribly.

Change-Id: I9ad46fca6b731a840578b11ba8119fc370af7c1c
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/39561
Reviewed-by: Daniel Erat <derat@chromium.org>
Reviewed-by: Sameer Nanda <snanda@chromium.org>