[init] fix date_proxy early exit
date_proxy was exiting early and in an unclean way, causing upstart to spawn it
again, which caused it to spawn new copies of its helper program, which would
then hang around forever. There are two causes of exits:

1. 'restart tlsdate' instead of 'restart tlsdated' causes us to exit because
this script is set +e, and
2. reading from the fifo returns eof when the writer end closes the fifo.

To solve 2, we have the writer end keep the fifo open for as long as it's
running by doing its logic in a subshell with stdout to the fifo.

BUG=chromium-os:36729
TEST=adhoc

Change-Id: I1275bee9f48aa342b669911e09f92c9eddd2e752
Signed-off-by: Elly Fong-Jones <ellyjones@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/39590
Reviewed-by: Kees Cook <keescook@chromium.org>