Make powerd_dbus_suspend wait synchronously on DBus resume signal
powerd_dbus_suspend is a one-liner shell script that sends a
RequestSuspend signal to powerd. It is currently only used in
suspend_stress_test, which needs high sleep delays and a bit of luck to
work, as the script returns immediately and does not block for the time
powerd needs to actually suspend the system.

This patch replaces that script with a small C++ binary that not only
sends the DBus signal, but also listens for a PowerStateChanged(on)
response from powerd_suspend (which is sent both on failure and on a
successful resume). This will make suspend_stress_test more reliable and
also enable autotests like power_UiResume to use powerd_dbus_suspend.

CQ-DEPEND=CL:39689
BUG=chromium-os:35901
TEST=Run powerd_dbus_suspend (while logged in), watch how the prompt
does not return until after the machine is resumed again.

Change-Id: I873c48de86591af5fb0c8e53c74e1ddcb69bc487
Signed-off-by: Julius Werner <jwerner@chromium.org>
Reviewed-on: https://gerrit.chromium.org/gerrit/39687