Revert "power: Wire up SuspendDelayController."
This reverts commit 268f0f3d6c61f5d19cc73b717da1366093d0b8ec

Original CL: https://gerrit.chromium.org/gerrit/#/c/39563/

This makes powerd's Suspender class pass suspend delay
method calls to SuspendDelayController and consult it to see
if it's safe to suspend.  The older implementation will be
preserved until all of the callers have been updated.

The suspend_delay_sample test program seemed to have a few
issues that prevented it from working even with the old
version of the suspend delay code, so this change fixes it
as well.  It uses the new implementation by default but has
a --old flag that can be set, along with --delay_ms to
control how long the program sleeps before reporting
readiness and --timeout_ms to set the timeout that will be
used when registering a delay.

BUG=chromium-os:36980
TEST=checked that suspend_delay_sample works both with and
     without --old.  observed powerd.LATEST and saw that
     suspends are deferred until all delays are accounted
     for.  checked that old-style delays used by existing
     programs still work.  checked that when "require
     password to wake from sleep" setting is enabled, powerd
     doesn't suspend until it additionally receives a
     ScreenIsUnlocked signal.  paused chrome in a debugger
     and checked that powerd suspends after three seconds of
     waiting for the screen to be locked.

Change-Id: I563847b679863ff493d300c5fbcc07e503be46b4
Reviewed-on: https://gerrit.chromium.org/gerrit/39706
Reviewed-by: David James <davidjames@chromium.org>
Commit-Ready: Thieu Le <thieule@chromium.org>
Tested-by: Thieu Le <thieule@chromium.org>